refactor: replace telemetry tracking functions with macros for improved reusability and cleanup unused session handlers
