+++ crates/goose-cli/src/cli/telemetry.rs
@@ -1,11 +1,7 @@
-use anyhow::Result;
 use std::collections::HashMap;
 
 use goose::message::MessageContent;
-use goose::telemetry::{
-    detect_environment, CommandExecution, CommandResult, CommandType, SessionExecution,
-    SessionResult, SessionType, ToolUsage, TelemetryExecution, SessionMetadataSupport,
-};
+use goose::telemetry::ToolUsage;
 
 pub fn log_if_telemetry_disabled(tracking_type: &str) {
     if goose::telemetry::global_telemetry().is_none() {
@@ -61,184 +57,4 @@ pub fn extract_tool_usage_from_session(session: &crate::Session) -> Vec<ToolUsag
     }
 
     tool_usage_map.into_values().collect()
-}
-
-pub async fn track_session_execution<F, Fut, T>(
-    session_id: &str,
-    session_type: SessionType,
-    execution_fn: F,
-) -> Result<T>
-where
-    F: FnOnce() -> Fut,
-    Fut: std::future::Future<Output = Result<(T, crate::Session)>>,
-{
-    let start_time = std::time::Instant::now();
-
-    let telemetry_execution = goose::telemetry::global_telemetry().map(|_manager| {
-        SessionExecution::new(session_id, session_type).with_metadata("interface", "cli")
-    });
-
-    log_if_telemetry_disabled("session");
-
-    let result = execution_fn().await;
-
-    if let Some(mut execution) = telemetry_execution {
-        let duration = start_time.elapsed();
-
-        match &result {
-            Ok((_, session)) => {
-                if let Ok(session_metadata) = session.get_metadata() {
-                    execution = execution.with_session_metadata(&session_metadata);
-                }
-
-                let tool_usage = extract_tool_usage_from_session(session);
-                for tool in tool_usage {
-                    execution.add_tool_usage(tool);
-                }
-
-                if let Some(env) = detect_environment() {
-                    execution = execution.with_environment(&env);
-                }
-
-                let messages = session.message_history();
-                execution = execution
-                    .with_turn_count(messages.len() as u64)
-                    .with_result(SessionResult::Success)
-                    .with_duration(duration);
-            }
-            Err(e) => {
-                execution = execution
-                    .with_result(SessionResult::Error(e.to_string()))
-                    .with_duration(duration);
-            }
-        }
-
-        if let Some(manager) = goose::telemetry::global_telemetry() {
-            if let Err(e) = manager.track_session_execution(execution).await {
-                tracing::warn!("Failed to track session execution: {}", e);
-            }
-        }
-    }
-
-    result.map(|(result, _)| result)
-}
-
-pub async fn track_command_execution<F, Fut, T>(
-    command_name: &str,
-    command_type: CommandType,
-    execution_fn: F,
-) -> Result<T>
-where
-    F: FnOnce() -> Fut,
-    Fut: std::future::Future<Output = Result<T>>,
-{
-    let start_time = std::time::Instant::now();
-
-    let telemetry_execution = goose::telemetry::global_telemetry()
-        .map(|_manager| CommandExecution::new(command_name, command_type));
-
-    log_if_telemetry_disabled("command");
-
-    let result = execution_fn().await;
-
-    if let Some(mut execution) = telemetry_execution {
-        let duration = start_time.elapsed();
-
-        match &result {
-            Ok(_) => {
-                execution = execution
-                    .with_result(CommandResult::Success)
-                    .with_duration(duration);
-            }
-            Err(e) => {
-                execution = execution
-                    .with_result(CommandResult::Error(e.to_string()))
-                    .with_duration(duration);
-            }
-        }
-
-        if let Some(env) = detect_environment() {
-            execution = execution.with_environment(&env);
-        }
-
-        if let Some(manager) = goose::telemetry::global_telemetry() {
-            if let Err(e) = manager.track_command_execution(execution).await {
-                tracing::warn!("Failed to track command execution: {}", e);
-            }
-        }
-    }
-
-    result
-}
-
-pub async fn track_recipe_execution<F, Fut, T>(
-    recipe_name: &str,
-    recipe_version: &str,
-    execution_fn: F,
-    params: Vec<(String, String)>,
-) -> Result<T>
-where
-    F: FnOnce() -> Fut,
-    Fut: std::future::Future<Output = Result<(T, crate::Session)>>,
-{
-    let start_time = std::time::Instant::now();
-
-    let telemetry_execution = goose::telemetry::global_telemetry()
-        .map(|manager| manager.recipe_execution(recipe_name, recipe_version));
-
-    log_if_telemetry_disabled("recipe");
-
-    let result = execution_fn().await;
-
-    if let Some(execution_builder) = telemetry_execution {
-        let duration = start_time.elapsed();
-        let execution = match &result {
-            Ok((_, session)) => {
-                let mut builder = execution_builder
-                    .with_result(goose::telemetry::SessionResult::Success)
-                    .with_duration(duration);
-
-                if let Ok(session_metadata) = session.get_metadata() {
-                    builder = builder.with_session_metadata(&session_metadata);
-                }
-
-                let tool_usage = extract_tool_usage_from_session(session);
-                for tool in tool_usage {
-                    builder = builder.add_tool_usage(tool);
-                }
-
-                for (key, value) in &params {
-                    builder = builder.with_metadata(key, value);
-                }
-
-                if let Some(env) = detect_environment() {
-                    builder = builder.with_environment(&env);
-                }
-
-                let messages = session.message_history();
-                builder = builder.with_turn_count(messages.len() as u64);
-
-                builder.build()
-            }
-            Err(e) => {
-                let mut builder = execution_builder
-                    .with_result(goose::telemetry::SessionResult::Error(e.to_string()))
-                    .with_duration(duration);
-
-                for (key, value) in &params {
-                    builder = builder.with_metadata(key, value);
-                }
-
-                builder.build()
-            }
-        };
-
-        if let Some(manager) = goose::telemetry::global_telemetry() {
-            if let Err(e) = manager.track_recipe_execution(execution).await {
-                tracing::warn!("Failed to track recipe execution: {}", e);
-            }
-        }
-    }
-
-    result.map(|(result, _)| result)
 }
